Restructure comp files & adjust imports
