+++ react-app/src/components/NotebooksView/DeleteNotebookView.js
