+++ react-app/src/components/SideMenu/SideMenu.css
@@ -20,4 +20,5 @@
 #nav-wrapper {
     display: flex;
     flex-direction: column;
+    overflow: auto;
 }
