+++ react-app/src/components/SideMenu/SideMenu.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { useDispatch } from 'react-redux';
 import { Redirect, useHistory } from 'react-router-dom';
-import { logout } from '../../../store/session';
+import { logout } from '../../store/session';
 import "./SideMenu.css";
 import NotebooksDropdown from "./NotebooksDropdown";
